+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Bug fixes:
 * Fixed `rb_get_kwargs` to correctly handle optional and rest arguments.
 * Calling `Kernel#raise` with a raised exception will no longer set the cause of the exception to itself (#1682).
 * Return a `FFI::Function` correctly for functions returning a callback.
+* Convert to intuitive Ruby exceptions when INVOKE fails (#1690).
 
 Compatibility