+++ CHANGELOG.md
@@ -17,6 +17,10 @@ Compatibility
 * `String#-@` now performs string deduplication (#1608).
 * `Hash#merge` now preserves the key order from the original hash for merged values (#1650).
 
+Changes:
+
+* Hash code calculation has been improved to reduce hash collisions for `Hash` and other cases.
+
 # 1.0 RC 17
 # 19.0.0, May 2019